// src/app/dashboard/pages/balances-page/interfaces/AllBalances.interface.ts
// Generated by https://quicktype.io

export interface AllBalances {
  notificacion: Notificacion;
  data:         Balance[];
}

export interface Balance {
  id:               number;
  forma_de_pago_id: number;
  ingresos:         string;
  egresos:          string;
  saldo:            string;
  created_at:       string;
  updated_at:       string;
  deleted_at:       null;
  forma_de_pago:    FormaDePago;
}

export interface FormaDePago {
  id:                   number;
  nombre:               string;
  tipo:                 number;
  nombre_titular_banco: string;
  banco:                string;
  cuenta:               string;
  clabe:                string;
  tarjeta:              string;
  usuario_id:           number;
  created_at:           string;
  updated_at:           string;
  deleted_at:           null;
}

export interface Notificacion {
  estado:  number;
  mensaje: string;
}