// src/app/dashboard/pages/balances-page/interfaces/Movements.interface.ts
// Generated by https://quicktype.io

export interface AllMovements {
  notificacion: Notificacion;
  data:         Movements[];
}

export interface Movements {
  tipomovimiento: string;
  movimiento:     Movimiento | null;
}

export interface Movimiento {
  id:               number;
  fecha:            string;
  proveedor_id?:    number;
  descripcion?:     string;
  importe:          string;
  forma_de_pago_id: number;
  usuario_id:       number;
  created_at:       string;
  updated_at:       string;
  deleted_at:       null;
  forma_de_pago:    FormaDePago;
  proveedor?:       Proveedor;
  usuario:          Usuario;
  pedido_id?:       number;
  pedido?:          Pedido;
}

export interface FormaDePago {
  id:                   number;
  nombre:               FormaDePagoNombre;
  tipo:                 number;
  nombre_titular_banco: string;
  banco:                string;
  cuenta:               string;
  clabe:                string;
  tarjeta:              string;
  usuario_id:           number;
  created_at:           string;
  updated_at:           string;
  deleted_at:           null;
}

export enum FormaDePagoNombre {
  EfectivoJonathan = "efectivo jonathan",
}

export interface Pedido {
  id:           number;
  cliente_id:   number;
  costo:        string;
  anticipo:     string;
  saldo:        string;
  usuario_id:   number;
  status:       number;
  created_at:   string;
  updated_at:   string;
  deleted_at:   null;
  fecha_pedido: string;
}

export interface Proveedor {
  id:         number;
  nombre:     string;
  direccion:  string;
  telefono:   string;
  email:      string;
  usuario_id: number;
  created_at: string;
  updated_at: string;
  deleted_at: null;
}


export interface Usuario {
  id:         number;
  nombre:     string;
  email:      string;
  password:   string;
  api_token:  string;
  created_at: string;
  updated_at: string;
  deleted_at: null;
}


export interface Notificacion {
  estado:  number;
  mensaje: string;
}